use Vector3f rather than Vector3d

+++ include/ray.h
@@ -7,11 +7,11 @@
 namespace zLi {
 
 struct Ray {
-  Vector3d o, d;
+  Vector3f o, d;
   Float tmin, tmax;
   Ray() {}
-  Ray(const Vector3d& o, const Vector3d& d): o(o), d(d), tmin(0), tmax(INF) {}
-  Ray(const Vector3d& o, const Vector3d& d, const Float tmin, const Float tmax): o(o), d(d), tmin(tmin), tmax(tmax) {}  
+  Ray(const Vector3f& o, const Vector3f& d): o(o), d(d), tmin(0), tmax(INF) {}
+  Ray(const Vector3f& o, const Vector3f& d, const Float tmin, const Float tmax): o(o), d(d), tmin(tmin), tmax(tmax) {}  
   Vector3f operator()(const Float t) const { return o + t * d; }
 };
